rozdzielenie metod
