+++ progress/WWL21221.csv
@@ -1,0 +1,1 @@
+Wyjazd,12.12.12,290000.0,1200.0
