+++ przejazd_proces.rb
@@ -1,16 +1,12 @@
 require 'csv'
 
 class Przejazd_proces
-    attr_accessor :numer_pojazdu, :data_wyjazdu, :km_wyjazd, :paliwo_dodatkowo, :paliwo_wyjazd, :data_zjazdu, :km_powrot, :paliwo_zjazd
-    def initialize(numer_pojazdu, data_wyjazdu, km_wyjazd, paliwo_wyjazd, paliwo_dodatkowo, data_zjazdu, km_powrot, paliwo_zjazd)
+    attr_accessor :numer_pojazdu, :data_wyjazdu, :km_wyjazd, :paliwo_dodatkowoot
+    def initialize(numer_pojazdu, data_wyjazdu, km_wyjazd, paliwo_wyjazd)
       @numer_pojazdu = numer_pojazdu
       @data_wyjazdu = data_wyjazdu
       @km_wyjazd = km_wyjazd
       @paliwo_wyjazd = paliwo_wyjazd
-      @paliwo_dodatkowo = paliwo_dodatkowo
-      @data_zjazdu = data_zjazdu
-      @km_powrot = km_powrot
-      @paliwo_zjazd = paliwo_zjazd
     end
 
     def wyjazd
