+++ zarejestruj_tankowanie.rb
@@ -1,0 +1,4 @@
+def cli_zarejestruj_dodatnkowanie
+    puts "Podaj ilość dotankowanego paliwa"
+    @paliwo_dodatkowo = gets.chomp.to_f
+  end
